// js/stat-evaluation-player/src/generated/FlickStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledCounts } from "./LabeledCounts.ts";

export type FlickStats = { count: number, high_confidence_count: number, is_last_flick: boolean, last_flick_time: number | null, last_flick_frame: number | null, time_since_last_flick: number | null, frames_since_last_flick: number | null, last_confidence: number | null, best_confidence: number, cumulative_confidence: number, cumulative_setup_duration: number, cumulative_ball_speed_change: number, labeled_event_counts?: LabeledCounts, };